requirements.txt
piper
# Core dependencies
piper-tts>=1.2.0
pygame>=2.5.0
pyaudio>=0.2.11
numpy>=1.24.0

# AI and ML dependencies
openai>=1.0.0
vosk>=0.3.45

# GPIO and hardware (Raspberry Pi)
RPi.GPIO>=0.7.1; sys_platform == "linux"

# Local LLM support
llama-cpp-python>=0.2.0

# Audio processing
soundfile>=0.12.1
scipy>=1.10.0

# Logging and utilities
python-dateutil>=2.8.2
requests>=2.31.0

# Development and testing
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
pytest-timeout>=2.1.0
pytest-xdist>=3.3.0
black>=23.0.0
flake8>=6.0.0
psutil>=5.9.0
playsound==1.2.2
